GridSprite class
